[WIP] GrapPA runs again, more debugging to do

+++ mlreco/iotools/samplers.py
@@ -80,7 +80,8 @@ def SequentialBatchSampler(distributed=False):
         def __iter__(self):
             """Iterates over sequential batches of data."""
             num_batches = self._num_samples/self._batch_size
-            order       = np.arange(num_batches*self._batch_size)
+            order       = np.arange(
+                    num_batches*self._batch_size, dtype=np.int64)
 
             return iter(order)
 
@@ -98,7 +99,9 @@ def RandomSequenceSampler(distributed=False):
             """
             max_id  = self._num_samples + 1 - self._batch_size
             starts  = self._random.randint(0, max_id, len(self))
-            batches = [np.arange(start, start+self._batch_size) for start in starts]
+            batches = [np.arange(
+                start, start + self._batch_size, 
+                dtype=np.int64) for start in starts]
 
             return iter(np.concatenate(batches))
 
@@ -118,9 +121,14 @@ def BootstrapBatchSampler(distributed=False):
             """Iterates over bootstrapped batches of data randomly picked
             from the dataset.
             """
-            starts = np.arange(0, self._num_samples+1 - self._batch_size, self._batch_size)
-            bootstrap_indices = np.random.choice(np.arange(self._num_samples), self._num_samples)
-            batches = [bootstrap_indices[np.arange(start, start+self._batch_size)] for start in starts]
+            starts = np.arange(
+                    0, self._num_samples+1 - self._batch_size, self._batch_size,
+                    dtype=np.int64)
+            bootstrap_indices = np.random.choice(np.arange(
+                self._num_samples), self._num_samples)
+            batches = [bootstrap_indices[np.arange(
+                start, start+self._batch_size,
+                dtype=np.int64)] for start in starts]
 
             return iter(np.concatenate(batches))
